--- a/crypto/fipsmodule/ml_kem/importer.sh
+++ b/crypto/fipsmodule/ml_kem/importer.sh
@@ -72,9 +72,17 @@ popd
 
 echo "Pull source code from remote repository..."
 
-# Copy mlkem-native source tree -- C-only, no FIPS-202
+# Copy mlkem-native source tree -- C source
 mkdir $SRC
 cp $TMP/mlkem/src/* $SRC
+
+# Copy AArch64 backend
+mkdir -p $SRC/native/aarch64/src
+# Backend API and specification assumed by mlkem-native frontend
+cp $TMP/mlkem/src/native/api.h $SRC/native
+# Copy AArch64 backend implementation
+cp $TMP/mlkem/src/native/aarch64/meta.h $SRC/native/aarch64
+cp $TMP/mlkem/src/native/aarch64/src/* $SRC/native/aarch64/src
 
 # We use the custom `mlkem_native_config.h`, so can remove the default one
 rm $SRC/config.h
@@ -84,11 +92,11 @@ cp $TMP/.clang-format $SRC
 
 # Copy and statically simplify BCM file
 # The static simplification is not necessary, but improves readability
-# by removing directives related to native backends that are irrelevant
-# for the C-only import.
+# by removing directives related to the FIPS-202 backend and the x86_64
+# arithmetic backend that are not yet imported.
 unifdef -DMLK_CONFIG_FIPS202_CUSTOM_HEADER                             \
-        -UMLK_CONFIG_USE_NATIVE_BACKEND_ARITH                          \
         -UMLK_CONFIG_USE_NATIVE_BACKEND_FIPS202                        \
+        -UMLK_SYS_X86_64                                               \
         $TMP/mlkem/mlkem_native.c                                      \
         > $SRC/mlkem_native_bcm.c
 
@@ -107,8 +115,29 @@ if [[ "$(uname)" == "Darwin" ]]; then
 else
   SED_I=(-i)
 fi
+
 echo "Fixup include paths"
 sed "${SED_I[@]}" 's/#include "src\/\([^"]*\)"/#include "\1"/' $SRC/mlkem_native_bcm.c
+
+echo "Fixup AArch64 assembly backend to use s2n-bignum macros"
+for file in $SRC/native/aarch64/src/*.S; do
+  echo "Processing $file"
+  tmp_file=$(mktemp)
+
+  # Flatten multiline preprocessor directives, then process with unifdef
+  sed -e ':a' -e 'N' -e '$!ba' -e 's/\\\n/ /g' "$file" | \
+  unifdef -DMLK_ARITH_BACKEND_AARCH64 -UMLK_CONFIG_MULTILEVEL_NO_SHARED -DMLK_CONFIG_MULTILEVEL_WITH_SHARED > "$tmp_file"
+  mv "$tmp_file" "$file"
+
+  # Replace common.h include and assembly macros
+  sed "${SED_I[@]}" 's/#include "\.\.\/\.\.\/\.\.\/common\.h"/#include "_internal_s2n_bignum.h"/' "$file"
+
+  func_name=$(grep -o '\.global MLK_ASM_NAMESPACE(\([^)]*\))' "$file" | sed 's/\.global MLK_ASM_NAMESPACE(\([^)]*\))/\1/')
+  if [ -n "$func_name" ]; then
+    sed "${SED_I[@]}" "s/\.global MLK_ASM_NAMESPACE($func_name)/        S2N_BN_SYM_VISIBILITY_DIRECTIVE(mlkem_$func_name)\n        S2N_BN_SYM_PRIVACY_DIRECTIVE(mlkem_$func_name)/" "$file"
+    sed "${SED_I[@]}" "s/MLK_ASM_FN_SYMBOL($func_name)/S2N_BN_SYMBOL(mlkem_$func_name):/" "$file"
+  fi
+done
 
 echo "Remove temporary artifacts ..."
 rm -rf $TMP