… crypto/fipsmodule/CMakeLists.txt
… crypto/fipsmodule/ml_kem/META.yml
… crypto/fipsmodule/ml_kem/mlkem/native/aarch64/src/aarch64_zetas.c
… crypto/fipsmodule/ml_kem/mlkem/native/aarch64/src/intt.S
… crypto/fipsmodule/ml_kem/mlkem/native/aarch64/src/ntt.S
… crypto/fipsmodule/ml_kem/mlkem/native/aarch64/src/polyvec_basemul_acc_montgomery_cached_asm_k2.S
… crypto/fipsmodule/ml_kem/mlkem/native/aarch64/src/polyvec_basemul_acc_montgomery_cached_asm_k3.S
… crypto/fipsmodule/ml_kem/mlkem/native/aarch64/src/polyvec_basemul_acc_montgomery_cached_asm_k4.S
… crypto/fipsmodule/ml_kem/mlkem/native/aarch64/src/rej_uniform_asm.S
… crypto/fipsmodule/ml_kem/mlkem/native/aarch64/src/rej_uniform_table.c
… crypto/fipsmodule/ml_kem/mlkem/native/api.h
… crypto/fipsmodule/ml_kem/mlkem_native_config.h