@hanno-becker 2025-07-18
@dkostic The backend is supposed to work with all AArch64 processors. Since Neon is mandatory with AArch64 and we do not use any extensions, no runtime checks are required (at least, as far as I know).